MAINT: Moved bg stuff to abc model. Small change to mpsa_eta
